# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BanHang Backend API
  version: 2.0.0
  description: Versioned NestJS backend API for auth, CRM, content, catalog, cart, orders, payments, inventory, shipping, notifications, reporting, and store locator flows.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email, password, and optional OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Auth payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/social/{provider}:
    post:
      tags: [Auth]
      summary: Social login or account linking with optional guest cart merge and referral tracking
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [google, facebook, apple, zalo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLoginRequest'
      responses:
        '200':
          description: Auth payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a customer account with optional guest cart merge and referral tracking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Auth payload with debug verification token outside production
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using body token or refresh cookie
      parameters:
        - in: header
          name: x-csrf-token
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed auth payload
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset requested
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified
  /auth/request-email-verification:
    post:
      tags: [Auth]
      summary: Request a new email verification token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification token requested
  /auth/2fa/enable:
    post:
      tags: [Auth]
      summary: Generate a TOTP secret and otpauth URL
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Two-factor setup payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFactorSetupResponse'
  /auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Verify TOTP code and activate two-factor authentication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyTwoFactorRequest'
      responses:
        '200':
          description: Two-factor verification result
  /auth/api-keys:
    get:
      tags: [Auth]
      summary: List API keys for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API key list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKeySummary'
    post:
      tags: [Auth]
      summary: Create an API key for integrations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
  /auth/api-keys/{id}:
    delete:
      tags: [Auth]
      summary: Revoke an API key
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key revoked
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
  /auth/sessions:
    get:
      tags: [Auth]
      summary: List refresh sessions for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session list
  /auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a refresh session
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session revoked
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke a refresh token or all current user sessions
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Logout success
...
  /compare:
    get:
      tags: [Compare]
      summary: List current compare products with differing field hints
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Compare list
  /compare/items:
    post:
      tags: [Compare]
      summary: Add a product into the compare list
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCompareItemRequest'
      responses:
        '200':
          description: Compare list updated
  /compare/items/{productId}:
    delete:
      tags: [Compare]
      summary: Remove one product from compare list
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Compare list updated
  /compare/clear:
    delete:
      tags: [Compare]
      summary: Clear the compare list
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Compare list cleared
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        otp:
          type: string
          minLength: 6
        guestAccessToken:
          type: string
          minLength: 20
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        guestAccessToken:
          type: string
          minLength: 20
        referralCode:
          type: string
          minLength: 4
...
    SocialLoginRequest:
      type: object
      required: [providerUserId, email]
      properties:
        providerUserId:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        phone:
          type: string
        guestAccessToken:
          type: string
          minLength: 20
        referralCode:
          type: string
          minLength: 4
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 40
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 40
        password:
          type: string
          minLength: 8
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 40
    VerifyTwoFactorRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 64
        permissions:
          type: array
          items:
            type: string
            enum: [auth.manage, catalog.read, catalog.write, cart.write, orders.read, orders.manage, payments.manage, inventory.read, inventory.manage, shipping.manage, notifications.read, reporting.read]
        expiresAt:
          type: string
          format: date-time
    TwoFactorSetupResponse:
      type: object
      properties:
        success:
          type: boolean
        enabled:
          type: boolean
        secret:
          type: string
        otpauthUrl:
          type: string
        debug:
          type: object
          properties:
            currentCode:
              type: string
    ApiKeySummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        keyPrefix:
          type: string
        permissions:
          type: array
          items:
            type: string
        lastUsedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ApiKeyCreateResponse:
      allOf:
        - $ref: '#/components/schemas/ApiKeySummary'
        - type: object
          properties:
            token:
              type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        csrfToken:
          type: string
        expiresIn:
          type: integer
        sessionId:
          type: string
        verificationRequired:
          type: boolean
        provider:
          type: string
        linkedAccount:
          type: boolean
        isNewAccount:
          type: boolean
        debug:
          type: object
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            fullName:
              type: string
            role:
              type: string
            phone:
              type: string
            referralCode:
              type: string
            emailVerifiedAt:
              type: string
              format: date-time
            twoFactorEnabledAt:
              type: string
              format: date-time
            lastLoginAt:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
...
    AddCompareItemRequest:
      type: object
      required: [productId]
      properties:
        productId:
          type: string
...